Put ElementComposite into a separate file
