+++ skfem/element/element.py
@@ -114,6 +114,18 @@ class Element():
         raise ValueError("Index larger than the number of basis functions.")
 
     def _bfun_counts(self, mapping):
+        """Count number of nodal/edge/facet/interior basis functions.
+
+        Parameters
+        ----------
+        mapping
+            :class:`skfem.mapping.Mapping`
+
+        Returns
+        -------
+        4-tuple of basis function counts.
+
+        """
         return (self.nodal_dofs * mapping.mesh.t.shape[0],
                 self.edge_dofs * mapping.mesh.t2e.shape[0]\
                 if hasattr(mapping.mesh, 'edges') else 0,
@@ -123,36 +135,11 @@ class Element():
 
     def __mul__(self, other):
 
-        class ElementComposite(Element):
+        from .element_composite import ElementComposite
 
-            def _deduce_bfun(self, mapping, i):
-                e1 = self.elems[0]._bfun_counts(mapping)
-                e2 = self.elems[1]._bfun_counts(mapping)
-                if i < e1[0]:
-                    return 0, i
-                elif i < e1[0] + e2[0]:
-                    return 1, i - e1[0]
-                elif i < e1[0] + e2[0] + e1[1]:
-                    return 0, i - e2[0]
-                elif i < e1[0] + e2[0] + e1[1] + e2[1]:
-                    return 1, i - e1[0] - e1[1]
-                elif i < e1[0] + e2[0] + e1[1] + e2[1] + e1[2]:
-                    return 0, i - e2[0] - e2[1]
-                elif i < e1[0] + e2[0] + e1[1] + e2[1] + e1[2] + e2[2]:
-                    return 1, i - e1[0] - e1[1] - e1[2]
-                elif i < e1[0] + e2[0] + e1[1] + e2[1] + e1[2] + e2[2] + e1[3]:
-                    return 0, i - e2[0] - e2[1] - e2[2]
-                return 1, i - e1[0] - e1[1] - e1[2] - e1[3]
-
-            def gbasis(self, mapping, X, i, **kwargs):
-                n, ind = self._deduce_bfun(mapping, i)
-                field = self.elems[n].gbasis(mapping, X, ind, **kwargs)[0]
-                if n == 0:
-                    Z = self.elems[1].gbasis(mapping, X, 0, **kwargs)[0].zeros_like()
-                    return (field, Z)
-                elif n == 1:
-                    Z = self.elems[0].gbasis(mapping, X, 0, **kwargs)[0].zeros_like()
-                    return (Z, field)
+        if isinstance(self, ElementComposite) or\
+           isinstance(other, ElementComposite):
+            raise NotImplementedError("Cannot combine two ElementComposite objects.")
 
         e = ElementComposite()
         e.elems = [self, other]
